# Conflicts:
#	src/main/kotlin/com/bossnutter/aoc2020/day07/Day07.kt

+++ src/main/kotlin/com/bossnutter/aoc2020/common/Util.kt
@@ -3,8 +3,8 @@ package com.bossnutter.aoc2020.common
 import java.io.File
 import java.util.stream.Collectors
 
-fun getNumbersFromFile(path: String): List<Int> =
-    File(path).readLines().stream().map { n -> Integer.valueOf(n) }.collect(Collectors.toList())
+fun getNumbersFromFile(path: String): List<Long> =
+    File(path).readLines().stream().map { n -> n.toLong() }.collect(Collectors.toList())
 
 fun <T> measureTimeMillisAndReturn(printFunction: (Long) -> Unit, function: () -> T): T {
     val startTime = System.currentTimeMillis()
